pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "clash-nodes"
version = "0.1.0"
description = "GitHub subscription discovery and subs-check pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9",
]

[project.scripts]
clash-nodes = "clash_nodes.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
